hashkit/crc16: fix conversion errors
Fixes compilation errors:
dyn_crc16.c:69:49: error: conversion to ‘uint32_t {aka unsigned int}’ from ‘char’ may change the sign of the result [-Werror=sign-conversion]
         crc = (crc << 8) ^ crc16tab[((crc >> 8) ^ *key++) & 0x00ff];

+++ src/hashkit/dyn_crc16.c
@@ -66,7 +66,7 @@ hash_crc16(const char *key, size_t key_length, struct dyn_token *token)
     uint32_t crc = 0;
 
     for (x=0; x < key_length; x++) {
-        crc = (crc << 8) ^ crc16tab[((crc >> 8) ^ *key++) & 0x00ff];
+        crc = (crc << 8) ^ crc16tab[((crc >> 8) ^ (size_t)*key++) & 0x00ff];
     }
 
     size_dyn_token(token, 1);
